requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.3.4
requests==2.32.3
reportlab==4.2.5
faiss-cpu==1.8.0.post1
pypdf==5.7.0
python-dotenv==1.0.1
email-validator==2.1.1
reportlab==4.2.0
structlog==24.4.0
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0
psutil==6.0.0
alembic==1.13.2
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
httpx==0.27.2
